remove dependencies to fedbiomed.node in fedbiomed.common
